Take into account a final `:`.
When passing in the follwing:
`org.scalameta:metals_2.12:`
the coursier completion will still think that you're at the artifact
stage and try to complete this. This is incorrect since the final `:`
is obviously indicating we want the versions. This change also
grabs that info and will correctly set the stage to VERSION if two
`:` are found with a valid org and artifact.